Move constants to shared
